Fix issue and implement previewing correct problem on results
